requirements.txt
mkdocs-material
PyYAML
requests
traceback-with-variables
typing_extensions
